+++ src/libs/logs/MarketplaceLogs.sol
@@ -3,13 +3,13 @@ pragma solidity 0.8.30;
 
 import {FeeType} from "@ticket-storage/MarketplaceStorage.sol";
 
-event TicketFeeSet(uint56 indexed ticketId, FeeType indexed feeType, uint256 fee);
+event TicketFeeSet(uint64 indexed ticketId, FeeType indexed feeType, uint256 fee);
 
 event TicketFeeAddressSet(FeeType indexed feeType, address indexed token);
 
-event TicketMinted(uint56 indexed ticketId, FeeType indexed feeType, uint256 fee, uint40 tokenId);
+event TicketMinted(uint64 indexed ticketId, FeeType indexed feeType, uint256 fee, uint40 tokenId);
 
-event TicketBalanceWithdrawn(uint56 indexed ticketId, FeeType indexed feeType, uint256 fee, address indexed to);
+event TicketBalanceWithdrawn(uint64 indexed ticketId, FeeType indexed feeType, uint256 fee, address indexed to);
 
 event HostItBalanceWithdrawn(FeeType indexed feeType, uint256 fee, address indexed to);
